rename : FrontConfig 파일 이동 - infra -> config
