+++ src/main/java/ssammudan/cotree/global/config/FrontendConfig.java
@@ -1,4 +1,4 @@
-package ssammudan.cotree.infra.frontend;
+package ssammudan.cotree.global.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
